+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ### Fixed
 - Fixed [#689](https://github.com/LykosAI/StabilityMatrix/issues/689) - New ComfyUI installs encountering launch error due to torch 2.0.0 update, added pinned `numpy==1.26.4` to install and update.
 - Fixed Inference image mask editor's 'Load Mask' not able to load image files
+- Fixed Fooocus ControlNet default config shared folder mode not taking effect
 
 ## v2.11.0
 ### Added